media-researcher-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "media-researcher-core"
version = "0.1.0"
description = "Core engine for media target discovery, enrichment, and scoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "diskcache>=5.6",
    "pydantic>=2.5",
    "python-dateutil>=2.9",
    "tenacity>=8.2",
    "notion-client>=2.2.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23", "respx>=0.21", "ruff>=0.4"]

[project.scripts]
media-researcher = "media_researcher_core.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["media_researcher_core"]

[tool.ruff]
line-length = 100
target-version = "py311"
